+++ store.config.js
@@ -52,7 +52,6 @@ module.exports = {
       home: '/',
       pdp: '/apple-magic-mouse/p',
       collection: '/office',
-      cmsHome: '/cms',
     },
   },
 
@@ -60,7 +59,6 @@ module.exports = {
   cypress: {
     pages: {
       home: '/',
-      cmsHome: '/cms',
       pdp: '/apple-magic-mouse/p',
       collection: '/office',
       collection_filtered:
